refactor run.rs into dedicated modules and files per command
